django: initial data: maintain lwm2m resouce model
The lwm2m resource model is defined in the specification. As of not
there was a .py mapping table that was mapping object/resource ids to a
name and data type. This can be made more elegant by directly adding it
to the database during first Django start. That way we can always
reference the database and no code changes are required for adding new
resources.

Signed-off-by: Jonas Remmert <jremmert@gmx.net>