Django: integrate Celery task queue to handle asynchronous tasks
Integrate Celery task queue. Redis serves as the message broker for the
Celery task queue. Redis can run in a seperate container with
of-the-shelf available images.

Gevent is used an efficient alternative compared to multiple python
processes, where each process has to load it's own runtime resulting in
high memory requirement. Celery is used for functions that wait for ReST
API calls to Leshan, an IO-bound task. Currently, Celery starts
1 worker (1 CPU), however, this worker is sufficient to handle more than
100 simultaneous ReST API calls efficiently.

Signed-off-by: Jonas Remmert <jremmert@gmx.net>