simulate: DL_SERVER: Add new KConfig symbol for DL Server URL
The download server for OTA updates could be different from the regular
server. The ota update does not send a complete URL, but only a path.
The host of the url is fixed in firmware.

This is a security feature to make sure the clients can never download
firmware from another than the specified server domain.

Signed-off-by: Jonas Remmert <jremmert@gmx.net>