django: implement timestamped serializer
When using the LWM2M SEND operation, leshan uses a different format,
called timestamped. In order to process SEND requests, this commit adds
a new timestamped serializer.

An Endpoint can send multiple resources with the SEND request at once.
In case multiple endpoints are send, django summarizes all of them
(regardless of the object) and links them to a new event. The name of
the event is the object of the first resource. In the example payload
below, the event name would be "3303".

Example SEND Operation:
{
  "ep" : "urn:imei:100000000000000",
  "val" : {
    "empty" : false,
    "timestamps" : [ null ],
    "nodes" : {
      "/3303/0/5700" : {
        "kind" : "singleResource",
        "id" : 5700,
        "type" : "FLOAT",
        "value" : "21.13306311567618"
      },
      "/3304/0/5700" : {
        "kind" : "singleResource",
        "id" : 5700,
        "type" : "FLOAT",
        "value" : "54.92426540118741"
      }
    }
  }
}

Signed-off-by: Jonas Remmert <jremmert@gmx.net>