SEO: increase site visibility
Increase site visibility by providing a robots.txt file that points to a
generated page sitemap.

In addition this commit adds a verification code to be able to analyze
it via Google Search Console. This does not track users or collect data,
but only analyzes how well the page can be indext and found.

Signed-off-by: Jonas Remmert <jremmert@gmx.net>